split up tests in one cpp file for each sorting network type to speed up compilation
